entrance_weights.c shared file extracted
