+++ config/splat.us.stnp3.yaml
@@ -166,7 +166,8 @@ segments:
       - [0x32830, c, entrance_stage_entities]
       - [0x32F30, c, e_castle_door]
       - [0x332A8, c, entrance_stage_entities_2]
-      - [0x344B4, c, entrance_stage_entities_3]
+      - [0x344B4, c, entrance_weights]
+      - [0x34BE8, c, entrance_stage_entities_3]
       - [0x35F60, c, e_heartroom]
       - [0x3644C, c, e_tilemap_shuffler]
       - [0x3653C, c, e_explosion_puff_opaque]
