+++ src/st/no3/entrance_weights.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "no3.h"
+
+#include "../entrance_weights.h"
